+++ src/test/java/com/influxdb/v3/client/internal/RestClientTest.java
@@ -39,7 +39,7 @@ import org.junit.jupiter.api.Test;
 
 import com.influxdb.v3.client.AbstractMockServerTest;
 import com.influxdb.v3.client.InfluxDBApiException;
-import com.influxdb.v3.client.config.InfluxDBClientConfigs;
+import com.influxdb.v3.client.config.ClientConfig;
 
 public class RestClientTest extends AbstractMockServerTest {
 
@@ -54,7 +54,7 @@ public class RestClientTest extends AbstractMockServerTest {
 
     @Test
     public void baseUrl() {
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder().hostUrl("http://localhost:8086").build());
+        restClient = new RestClient(new ClientConfig.Builder().host("http://localhost:8086").build());
         Assertions
                 .assertThat(restClient.baseUrl)
                 .isEqualTo("http://localhost:8086/");
@@ -62,7 +62,7 @@ public class RestClientTest extends AbstractMockServerTest {
 
     @Test
     public void baseUrlSlashEnd() {
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder().hostUrl("http://localhost:8086/").build());
+        restClient = new RestClient(new ClientConfig.Builder().host("http://localhost:8086/").build());
         Assertions
                 .assertThat(restClient.baseUrl)
                 .isEqualTo("http://localhost:8086/");
@@ -70,9 +70,9 @@ public class RestClientTest extends AbstractMockServerTest {
 
     @Test
     public void responseTimeout() {
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl("http://localhost:8086")
-                .responseTimeout(Duration.ofSeconds(13))
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host("http://localhost:8086")
+                .timeout(Duration.ofSeconds(13))
                 .build());
 
         Optional<Duration> connectTimeout = restClient.client.connectTimeout();
@@ -83,8 +83,8 @@ public class RestClientTest extends AbstractMockServerTest {
 
     @Test
     public void allowHttpRedirectsDefaults() {
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl("http://localhost:8086")
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host("http://localhost:8086")
                 .build());
 
         HttpClient.Redirect redirect = restClient.client.followRedirects();
@@ -95,9 +95,9 @@ public class RestClientTest extends AbstractMockServerTest {
     public void authenticationHeader() throws InterruptedException {
         mockServer.enqueue(createResponse(200));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl(baseURL)
-                .authToken("my-token".toCharArray())
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host(baseURL)
+                .token("my-token".toCharArray())
                 .build());
 
         restClient.request("ping", HttpMethod.GET, null, null, null);
@@ -112,8 +112,8 @@ public class RestClientTest extends AbstractMockServerTest {
     public void authenticationHeaderNotDefined() throws InterruptedException {
         mockServer.enqueue(createResponse(200));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl(baseURL)
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host(baseURL)
                 .build());
 
         restClient.request("ping", HttpMethod.GET, null, null, null);
@@ -128,8 +128,8 @@ public class RestClientTest extends AbstractMockServerTest {
     public void userAgent() throws InterruptedException {
         mockServer.enqueue(createResponse(200));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl(baseURL)
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host(baseURL)
                 .build());
 
         restClient.request("ping", HttpMethod.GET, null, null, null);
@@ -144,9 +144,9 @@ public class RestClientTest extends AbstractMockServerTest {
     public void customHeader() throws InterruptedException {
         mockServer.enqueue(createResponse(200));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl(baseURL)
-                .authToken("my-token".toCharArray())
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host(baseURL)
+                .token("my-token".toCharArray())
                 .headers(Map.of("X-device", "ab-01"))
                 .build());
 
@@ -162,8 +162,8 @@ public class RestClientTest extends AbstractMockServerTest {
     public void uri() throws InterruptedException {
         mockServer.enqueue(createResponse(200));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl(baseURL)
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host(baseURL)
                 .build());
 
         restClient.request("ping", HttpMethod.GET, null, null, null);
@@ -176,8 +176,8 @@ public class RestClientTest extends AbstractMockServerTest {
 
     @Test
     public void allowHttpRedirects() {
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl("http://localhost:8086")
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host("http://localhost:8086")
                 .allowHttpRedirects(true)
                 .build());
 
@@ -189,8 +189,8 @@ public class RestClientTest extends AbstractMockServerTest {
     public void proxy() throws InterruptedException {
         mockServer.enqueue(createResponse(200));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl("http://foo.com:8086")
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host("http://foo.com:8086")
                 .proxy(ProxySelector.of((InetSocketAddress) mockServer.toProxyAddress().address()))
                 .build());
 
@@ -208,8 +208,8 @@ public class RestClientTest extends AbstractMockServerTest {
         mockServer.enqueue(createResponseWithHeaders(407, Map.of("Proxy-Authenticate", "Basic")));
         mockServer.enqueue(createResponse(200));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl("http://foo.com:8086")
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host("http://foo.com:8086")
                 .proxy(ProxySelector.of((InetSocketAddress) mockServer.toProxyAddress().address()))
                 .authenticator(new Authenticator() {
                     @Override
@@ -238,8 +238,8 @@ public class RestClientTest extends AbstractMockServerTest {
     public void error() {
         mockServer.enqueue(createResponse(404));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl(baseURL)
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host(baseURL)
                 .build());
 
         Assertions.assertThatThrownBy(
@@ -252,8 +252,8 @@ public class RestClientTest extends AbstractMockServerTest {
     public void errorFromHeader() {
         mockServer.enqueue(createResponse(500).setHeader("X-Influx-Error", "not used"));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl(baseURL)
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host(baseURL)
                 .build());
 
         Assertions.assertThatThrownBy(
@@ -268,8 +268,8 @@ public class RestClientTest extends AbstractMockServerTest {
                 .setHeader("X-Influx-Error", "not used")
                 .setBody("{\"message\":\"token does not have sufficient permissions\"}"));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl(baseURL)
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host(baseURL)
                 .build());
 
         Assertions.assertThatThrownBy(
@@ -283,8 +283,8 @@ public class RestClientTest extends AbstractMockServerTest {
         mockServer.enqueue(createResponse(402)
                 .setBody("token is over the limit"));
 
-        restClient = new RestClient(new InfluxDBClientConfigs.Builder()
-                .hostUrl(baseURL)
+        restClient = new RestClient(new ClientConfig.Builder()
+                .host(baseURL)
                 .build());
 
         Assertions.assertThatThrownBy(